Modular refactoring and unit tests to showcase usage
