+++ nemo/collections/asr/models/hybrid_rnnt_ctc_bpe_models.py
@@ -129,17 +129,20 @@ class EncDecHybridRNNTCTCBPEModel(EncDecHybridRNNTCTCModel, ASRBPEMixin):
     def _setup_dataloader_from_config(self, config: Optional[Dict]):
 
         if config.get("use_lhotse"):
-            from nemo.collections.asr.data.audio_to_text_lhotse import (
-                get_lhotse_audio_to_text_char_dataloader_from_config,
-            )
+            from nemo.collections.asr.data.audio_to_text_lhotse import LhotseSpeechToTextBpeDataset
+            from nemo.collections.asr.data.lhotse.dataloader import get_lhotse_dataloader_from_config
 
-            return get_lhotse_audio_to_text_char_dataloader_from_config(
+            # Note:
+            #    Lhotse Dataset only maps CutSet -> batch of tensors, but does not actually
+            #    contain any data or meta-data; it is passed to it by a Lhotse sampler for
+            #    each sampler mini-batch.
+            return get_lhotse_dataloader_from_config(
                 config,
-                local_rank=self.local_rank,
                 global_rank=self.global_rank,
                 world_size=self.world_size,
-                tokenizer=self.tokenizer,
-                preprocessor_cfg=self.cfg.get("precprocessor", None),
+                dataset=LhotseSpeechToTextBpeDataset(
+                    tokenizer=self.tokenizer, noise_cuts=config.get("lhotse", {}).get("noise_cuts")
+                ),
             )
 
         dataset = audio_to_text_dataset.get_audio_to_text_bpe_dataset_from_config(
