+++ tests/collections/asr/test_lhotse_dataloading.py
@@ -1,0 +1,274 @@
+import tarfile
+from io import BytesIO
+from itertools import islice
+from pathlib import Path
+from typing import Dict, Tuple, List
+
+import pytest
+import torch.utils.data
+from omegaconf import OmegaConf
+
+from nemo.collections.asr.data.lhotse.dataloader import get_lhotse_dataloader_from_config
+
+lhotse = pytest.importorskip(
+    "lhotse", reason="Lhotse + NeMo tests require Lhotse to be installed (pip install lhotse)."
+)
+
+
+@pytest.fixture()
+def cutset_path(tmp_path: Path) -> Path:
+    """10 utterances of length 1s as a Lhotse CutSet."""
+    from lhotse.testing.dummies import DummyManifest
+    from lhotse import CutSet
+
+    cuts = DummyManifest(CutSet, begin_id=0, end_id=10, with_data=True)
+    for c in cuts:
+        c.features = None
+        c.custom = None
+        c.supervisions[0].custom = None
+    p = tmp_path / "cuts.jsonl.gz"
+    pa = tmp_path / "audio"
+    cuts.save_audios(pa).to_file(p)
+    return p
+
+
+@pytest.fixture()
+def cutset_shar_path(cutset_path: Path) -> Path:
+    """10 utterances of length 1s as a Lhotse Shar (tarred) CutSet."""
+    from lhotse import CutSet
+
+    cuts = CutSet.from_file(cutset_path)
+    p = cutset_path.parent / "shar"
+    p.mkdir(exist_ok=True)
+    cuts.to_shar(p, fields={"recording": "wav"}, shard_size=5)
+    return p
+
+
+@pytest.fixture()
+def nemo_manifest_path(cutset_path: Path):
+    """10 utterances of length 1s as a NeMo manifest."""
+    from lhotse import CutSet
+    from lhotse.serialization import save_to_jsonl
+
+    nemo = []
+    for c in CutSet.from_file(cutset_path):
+        nemo.append({"audio_filepath": c.recording.sources[0].source, "text": "irrelevant", "duration": c.duration})
+    p = cutset_path.parent / "nemo_manifest.json"
+    save_to_jsonl(nemo, p)
+    return p
+
+
+@pytest.fixture()
+def nemo_tarred_manifest_path(nemo_manifest_path: Path) -> Tuple[str, str]:
+    """10 utterances of length 1s as a NeMo tarred manifest."""
+    from lhotse.serialization import load_jsonl, SequentialJsonlWriter
+    from lhotse.shar.writers import TarWriter
+
+    root = nemo_manifest_path.parent / "nemo_tar"
+    root.mkdir(exist_ok=True)
+
+    with TarWriter(f"{root}/audios_%01d.tar", shard_size=5) as tar_writer, SequentialJsonlWriter(
+        root / "tarred_audio_filepaths.jsonl"
+    ) as mft_writer:
+        for idx, d in enumerate(load_jsonl(nemo_manifest_path)):
+            p = d["audio_filepath"]
+            tar_writer.write(p, BytesIO(open(p, "rb").read()))
+            mft_writer.write({**d, "audio_filepath": Path(p).name, "shard_id": int(idx > 4)})
+    return mft_writer.path, f"{root}/audios__OP_0..1_CL_.tar"
+
+
+class UnsupervisedAudioDataset(torch.utils.data.Dataset):
+    def __getitem__(self, cuts: lhotse.CutSet) -> Dict[str, torch.Tensor]:
+        audio, audio_lens = lhotse.dataset.collation.collate_audio(cuts)
+        return {"audio": audio, "audio_lens": audio_lens}
+
+
+def test_dataloader_from_lhotse_cuts(cutset_path: Path):
+    config = OmegaConf.create(
+        {
+            "sample_rate": 16000,
+            "shuffle": True,
+            "use_lhotse": True,
+            "num_workers": 0,
+            "lhotse": {
+                "cuts_path": cutset_path,
+                "use_bucketing": True,
+                "num_buckets": 2,
+                "drop_last": False,
+                "batch_duration": 4.0,  # seconds
+                "quadratic_duration": 15.0,  # seconds
+                "seed": 0,
+            },
+        }
+    )
+
+    dl = get_lhotse_dataloader_from_config(
+        config=config, global_rank=0, world_size=1, dataset=UnsupervisedAudioDataset()
+    )
+
+    batches = [batch for batch in dl]
+    assert len(batches) == 4
+
+    b = batches[0]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[1]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[2]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[3]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 1
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+
+def test_dataloader_from_lhotse_shar_cuts(cutset_shar_path: Path):
+    config = OmegaConf.create(
+        {
+            "sample_rate": 16000,
+            "shuffle": True,
+            "use_lhotse": True,
+            "num_workers": 0,
+            "lhotse": {
+                "shar_path": cutset_shar_path,
+                "use_bucketing": True,
+                "num_buckets": 2,
+                "drop_last": False,
+                "batch_duration": 4.0,  # seconds
+                "quadratic_duration": 15.0,  # seconds
+                "seed": 0,
+            },
+        }
+    )
+
+    dl = get_lhotse_dataloader_from_config(
+        config=config, global_rank=0, world_size=1, dataset=UnsupervisedAudioDataset()
+    )
+
+    # Note: we use islice here because with Lhotse Shar the dataloader will always be infinite.
+    batches = [batch for batch in islice(dl, 4)]
+    assert len(batches) == 4
+
+    b = batches[0]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[1]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[2]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[3]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+
+def test_dataloader_from_nemo_manifest(nemo_manifest_path: Path):
+    config = OmegaConf.create(
+        {
+            "manifest_filepath": nemo_manifest_path,
+            "sample_rate": 16000,
+            "shuffle": True,
+            "use_lhotse": True,
+            "num_workers": 0,
+            "lhotse": {
+                "use_bucketing": True,
+                "num_buckets": 2,
+                "drop_last": False,
+                "batch_duration": 4.0,  # seconds
+                "quadratic_duration": 15.0,  # seconds
+                "seed": 0,
+            },
+        }
+    )
+
+    dl = get_lhotse_dataloader_from_config(
+        config=config, global_rank=0, world_size=1, dataset=UnsupervisedAudioDataset()
+    )
+
+    batches = [batch for batch in dl]
+    assert len(batches) == 4
+
+    b = batches[0]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[1]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[2]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[3]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 1
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+
+def test_dataloader_from_tarred_nemo_manifest(nemo_tarred_manifest_path: Path):
+    json_mft, tar_mft = nemo_tarred_manifest_path
+    config = OmegaConf.create(
+        {
+            "manifest_filepath": json_mft,
+            "tarred_audio_filepaths": tar_mft,
+            "sample_rate": 16000,
+            "shuffle": True,
+            "use_lhotse": True,
+            "num_workers": 0,
+            "lhotse": {
+                "use_bucketing": True,
+                "num_buckets": 2,
+                "drop_last": False,
+                "batch_duration": 4.0,  # seconds
+                "quadratic_duration": 15.0,  # seconds
+                "seed": 0,
+            },
+        }
+    )
+
+    dl = get_lhotse_dataloader_from_config(
+        config=config, global_rank=0, world_size=1, dataset=UnsupervisedAudioDataset()
+    )
+
+    batches = [batch for batch in dl]
+    assert len(batches) == 4
+
+    b = batches[0]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[1]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[2]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 3
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
+
+    b = batches[3]
+    assert set(b.keys()) == {"audio", "audio_lens"}
+    assert b["audio"].shape[0] == b["audio_lens"].shape[0] == 1
+    assert all(isinstance(v, torch.Tensor) for v in b.values())
